feat(jobs): Job Kernel + JobType 插件化重构 —— Job 只做任务索引与生命周期外壳
- 新增 jobs/job_types：JobTypeRegistry + JobTypeHandler，xTB 优化 / Psi4 频率 / TS / IRC 注册为四个版本化 JobType，各带独立 request/executor/collector，worker 经 Registry 分发不再按任务类型写分支
- 新增中立引擎层 engines/xtb（契约/几何 IO/命令构造/输出解析/进程执行），/optimize 同步接口与持久 Job 共用 run_xtb_process，routers/optimize.py 瘦身 ~500 行
- Schema 10：JobTypeData / JobRun / Artifact.runId；worker 领取时建 Run，(run_id, artifact.name) 幂等键；只读 API /jobs/{id}/type-data、/runs、/runs/{runId}
- 创建阶段由 JobType 规范化参数后才入库；worker 分发前校验 jobType@version 与注册实现一致
- 边界测试拒绝 job_types → routers 反向导入；架构决策文档 job-kernel-and-job-types.md

+++ software/backend/engines/process_runner.py
@@ -27,7 +27,7 @@ def run_live_process(
     *,
     cwd: Path,
     log_path: Path,
-    timeout: int,
+    timeout: float,
     cancel_check: Callable[[], bool] | None = None,
     poll_interval: float = 0.2,
 ) -> LiveProcessResult:
